+++ src/kanibako/commands/code_cmd.py
@@ -194,7 +194,7 @@ def run_code(args: argparse.Namespace) -> int:
     except ProjectError:
         proj = None
     if proj is None or not proj.name:
-        print(_no_box_error(project_dir), file=sys.stderr)
+        print(_no_box_error(project_dir, std), file=sys.stderr)
         return 1
     cname = container_name_for(proj)